requirements.txt
Flask
Flask-Cors  # Capitalization matters!
openai
requests
python-dotenv
